software/pyproject.toml
[project]
name = "NeuroSync"
version = "0.1.0"
requires-python = "~=3.13.0"

dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "tabulate",
    "pylsl",
    "matplotlib",
    "crcmod>=1.7",
    "pytest>=9.0.2",
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "black>=26.1.0",
    "docformatter>=1.7.7",
    "flake8>=7.3.0",
    "flake8-pyproject>=1.2.4",
    "isort>=7.0.0",
    "pytest-cov>=7.1.0",
    "safety>=3.7.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[[tool.uv.index]]
url = "https://pkgs.safetycli.com/repository/public/pypi/simple/"
default = true

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.isort]
profile = "black"
line_length = 79
skip = [".venv", ".git", "__pycache__"]

[tool.black]
line-length = 79
target-version = ["py313"]


[tool.pyright]
ignore = ["**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

